+++ be/src/main/java/kr/codesquad/jazzmeet/venue/controller/annotation/Longitude.java
@@ -1,4 +1,4 @@
-package kr.codesquad.jazzmeet.venue.dto.annotation;
+package kr.codesquad.jazzmeet.venue.controller.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -8,12 +8,12 @@ import java.lang.annotation.Target;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 
-@Target(ElementType.TYPE)
+@Target({ElementType.PARAMETER, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = RangeCoordinatesValidator.class)
-public @interface RangeCoordinates {
+@Constraint(validatedBy = LongitudeValidator.class)
+public @interface Longitude {
 
-	String message() default "좌표의 범위가 유효하지 않습니다.";
+	String message() default "경도의 범위가 유효하지 않습니다.";
 
 	Class<?>[] groups() default {};
