+++ be/src/main/java/kr/codesquad/jazzmeet/venue/controller/annotation/RangeCoordinates.java
@@ -1,4 +1,4 @@
-package kr.codesquad.jazzmeet.venue.dto.annotation;
+package kr.codesquad.jazzmeet.venue.controller.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -8,12 +8,12 @@ import java.lang.annotation.Target;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 
-@Target({ElementType.PARAMETER, ElementType.FIELD})
+@Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = LongitudeValidator.class)
-public @interface Longitude {
+@Constraint(validatedBy = RangeCoordinatesValidator.class)
+public @interface RangeCoordinates {
 
-	String message() default "경도의 범위가 유효하지 않습니다.";
+	String message() default "좌표의 범위가 유효하지 않습니다.";
 
 	Class<?>[] groups() default {};
