+++ be/src/main/java/kr/codesquad/jazzmeet/venue/controller/annotation/RangeCoordinatesValidator.java
@@ -1,4 +1,4 @@
-package kr.codesquad.jazzmeet.venue.dto.annotation;
+package kr.codesquad.jazzmeet.venue.controller.annotation;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
